pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "golf-scorecards"
version = "0.1.0"
description = "Golf scorecard and performance metrics backend."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.0,<1.0.0",
    "jinja2>=3.1.0,<4.0.0",
    "pydantic-settings>=2.11.0,<3.0.0",
    "python-multipart>=0.0.20,<1.0.0",
    "uvicorn[standard]>=0.35.0,<1.0.0",
    "weasyprint>=63.0,<64.0",
]

[project.scripts]
golf-scorecards = "golf_scorecards.cli:main"

[dependency-groups]
dev = [
    "httpx>=0.28.0,<1.0.0",
    "mypy>=1.18.0,<2.0.0",
    "pytest>=8.4.0,<9.0.0",
    "pytest-cov>=7.0.0,<8.0.0",
    "ruff>=0.13.0,<1.0.0",
]

[tool.hatch.build]
include = [
    "src/golf_scorecards/data/**/*.json",
    "src/golf_scorecards/static/**/*",
    "src/golf_scorecards/templates/**/*",
]

[tool.hatch.build.targets.wheel]
packages = ["src/golf_scorecards"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.lint.per-file-ignores]
"src/golf_scorecards/web/routes.py" = ["B008"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
mypy_path = ["src"]
explicit_package_bases = true
plugins = []
